+++ core-ng/src/main/java/core/framework/http/HTTPHeaders.java
@@ -6,7 +6,7 @@ package core.framework.http;
 public final class HTTPHeaders {
     // http header is case insensitive, use lower case value to match jdk and chrome style
     public static final String ACCEPT = "accept";
-    public static final String ACCEPT_ENCODING = "accept-encoding";
+    public static final String ACCEPT_ENCODING = "accept-encoding"; // TODO: not used anymore if switch to okhttp
     public static final String AUTHORIZATION = "authorization";
     public static final String CACHE_CONTROL = "cache-control";
     public static final String CONTENT_TYPE = "content-type";